UX-675 convert useTabs to typescript
