+++ packages/matchbox/src/components/Tabs/index.ts
@@ -1,1 +1,3 @@
 export { default as Tabs } from './Tabs';
+export * from './Tabs';
+export * from './Tab';
